refactor: 불필요한 파일 삭제 및 클래스명 변경
