--- src/main/java/chess/utils/SQLConnection.java
@@ -1,27 +1,0 @@
-package chess.utils;
-
-import chess.exception.DbConnectionException;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-public class SQLConnection {
-
-    private static final String DRIVER = "com.mysql.cj.jdbc.Driver";
-    private static final String URL = "jdbc:mysql://localhost:3306/chess";
-    private static final String USER = "root";
-    private static final String PASSWORD = "root";
-
-
-    public static Connection getConnection() {
-        Connection connection = null;
-        try {
-            Class.forName(DRIVER);
-            connection = DriverManager.getConnection(URL, USER, PASSWORD);
-        } catch (ClassNotFoundException | SQLException e) {
-            e.printStackTrace();
-            throw new DbConnectionException();
-        }
-        return connection;
-    }
-}
